refactor: add business logic tier
