+++ src/controllers/transactionController.ts
@@ -1,7 +1,6 @@
-import {TransactionType} from "../models/transactionModels";
-
 const Transaction = require("../models/transactionModels")
 import {Request, Response} from "express";
+import {getLimitedHistoryAndSummary} from "../services/transactionService";
 
 exports.getTransactionHistory = async (req: Request, res: Response) => {
     try {
@@ -59,24 +58,9 @@ exports.updateTransaction = async (req: Request, res: Response) => {
 }
 
 exports.getTransactionsSummary = async (req: Request, res: Response) => {
-    const getMoneyGroupedByPerson = (history: Array<TransactionType>) => {
-        const summary: { [borrowedBy: string]: number } = {}
-        history.forEach((transaction: TransactionType) => {
-            if (summary[transaction.borrowedBy] == null) {
-                summary[transaction.borrowedBy] = 0
-            }
-            summary[transaction.borrowedBy] += transaction.amount
-        })
-        return summary
-    }
-
     try {
-        const historyListLength: number | undefined = (req.query.historyListLength && Number.parseInt(req.query.historyListLength as string)) || undefined
-        const history = await Transaction.find().sort({'timestamp': -1})
-        const summary: { [borrowedBy: string]: number } = getMoneyGroupedByPerson(history)
-        const limitedHistory = (historyListLength && historyListLength - 1 < history.length) ?
-            history.slice(0, historyListLength) : history
-
+        // TODO: jeszcze jest error
+        const {summary, limitedHistory} = getLimitedHistoryAndSummary(req)
         res.status(200).json({
             status: 'success',
             results: Object.keys(summary).length,
